+++ quill/include/quill/core/LoggerBase.h
@@ -25,6 +25,7 @@ class UserClockSource;
 namespace detail
 {
 class BackendWorker;
+class BacktraceStorage;
 
 /***/
 class LoggerBase
@@ -34,12 +35,12 @@ public:
   LoggerBase(std::string logger_name, std::vector<std::shared_ptr<Sink>> sinks,
              std::string format_pattern, std::string time_pattern, Timezone timezone,
              ClockSourceType clock_source, UserClockSource* user_clock)
-    : format_pattern(static_cast<std::string&&>(format_pattern)),
-      time_pattern(static_cast<std::string&&>(time_pattern)),
-      logger_name(static_cast<std::string&&>(logger_name)),
+    : clock_source(clock_source),
       user_clock(user_clock),
+      logger_name(static_cast<std::string&&>(logger_name)),
       timezone(timezone),
-      clock_source(clock_source)
+      format_pattern(static_cast<std::string&&>(format_pattern)),
+      time_pattern(static_cast<std::string&&>(time_pattern))
   {
 #ifndef NDEBUG
     for (auto const& sink : sinks)
@@ -123,18 +124,25 @@ public:
 protected:
   friend class detail::BackendWorker;
 
-  std::shared_ptr<PatternFormatter> pattern_formatter; /* The backend thread will set this once, we never access it on the frontend */
-  std::string format_pattern; /* Set by the frontend and accessed by the backend to initialise PatternFormatter */
-  std::string time_pattern; /* Set by the frontend and accessed by the backend to initialise PatternFormatter */
-  std::string logger_name; /* Set by the frontend, accessed by the frontend AND backend */
-  std::vector<std::shared_ptr<Sink>> sinks; /* Set by the frontend and accessed by the backend */
-  UserClockSource* user_clock{nullptr}; /* A non owned pointer to a custom timestamp clock, valid only when provided. used by frontend only */
-  Timezone timezone; /* Set by the frontend and accessed by the backend to initialise PatternFormatter */
-  ClockSourceType clock_source; /* Set by the frontend and accessed by the frontend AND backend */
+  /** Accessed by Frontend frequently - hot **/
+  std::atomic<LogLevel> log_level{LogLevel::Info}; /* Accessed by frontend only */
+  ClockSourceType clock_source; /* Accessed by the frontend, accessed but not modified by the frontend AND backend */
+  UserClockSource* user_clock{nullptr}; /* A non owned pointer to a custom timestamp clock, valid only when provided. accessed by frontend only */
+  std::atomic<LogLevel> backtrace_flush_level{LogLevel::None}; /** Modified by the frontend, accessed but not modified by the backend */
 
-  std::atomic<LogLevel> log_level{LogLevel::Info}; /* used by frontend only */
-  std::atomic<LogLevel> backtrace_flush_level{LogLevel::None}; /** Updated by the frontend at any time, accessed by the backend */
-  std::atomic<bool> valid{true}; /* Updated by the frontend at any time, accessed by the backend */
+  /** Backend only variables - Modified by the Backend **/
+  alignas(CACHE_LINE_ALIGNED) std::shared_ptr<PatternFormatter> pattern_formatter; /* The backend thread will construct this, we never access it on the frontend */
+  std::shared_ptr<BacktraceStorage> backtrace_storage; /* The backend thread will construct this, we never access it on the frontend */
+
+  /** Frequent Backend access - Rare Frontend access **/
+  std::string logger_name; /* Set by the frontend, accessed rarely by the frontend */
+  std::vector<std::shared_ptr<Sink>> sinks; /* Set by the frontend */
+  std::atomic<bool> valid{true}; /* Modified by the frontend, accessed but not modified the backend */
+
+  /** Set by the Frontend on constructor - Backend access once **/
+  Timezone timezone; /* Set by the frontend and accessed by the backend to construct PatternFormatter */
+  std::string format_pattern; /* Set by the frontend and accessed by the backend to construct PatternFormatter */
+  std::string time_pattern; /* Set by the frontend and accessed by the backend to construct PatternFormatter */
 };
 } // namespace detail
 } // namespace quill